AI/requirements.txt
openai
fastapi
uvicorn
python-dotenv
httpx
pydantic
python-multipart